requirements.txt
Flask==3.0.0
google-cloud-storage==2.10.0
appwrite==4.1.0
python-dotenv==1.0.0
together==0.2.7
PyMuPDF==1.23.8
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
playwright==1.40.0
requests==2.31.0